split tags.vue into tags.vue and TagFeed.vue. Split testing also
